feat(remix): Conditionally render admin setup form
